change type to long

+++ src/main/java/lab3/API.java
@@ -23,14 +23,14 @@ public class API {
         System.out.println(myObject);
     }
 
-    public int play(String id, String mode, int bet, long number) throws IOException, InterruptedException {
+    public long play(String id, String mode, int bet, long number) throws IOException, InterruptedException {
         HttpClient client = HttpClient.newHttpClient();
         HttpRequest request = HttpRequest.newBuilder().uri(URI
                 .create(serverURL + "/play" + mode + "?id=" + id + "&bet=" +
                         bet + "&number=" + number)).build();
         HttpResponse<String> response = client.send(request, HttpResponse.BodyHandlers.ofString());
         JSONObject myObject = new JSONObject(response.body());
-        int realNumber = myObject.getInt("realNumber");
+        long realNumber = myObject.getLong("realNumber");
         System.out.println(myObject);
         return realNumber;
     }
